Adicionado Login com token e restrições no banco de dados

+++ apps/backend-c#/DTO/TournamentDTO.cs
@@ -4,7 +4,7 @@ public class TournamentDTO
   public int tournamentId { get; set; }
   public string tournamentName { get; set; } = String.Empty;
   public string tournamentStatus { get; set; } = String.Empty;
-  public int prizeMoney { get; set; }
+  public decimal prizeMoney { get; set; }
   public string admFirstName { get; set; } = String.Empty;
   public string admLastName { get; set; } = String.Empty;
 
@@ -15,7 +15,7 @@ public class TournamentAddDTO
   public int tournamentId { get; set; }
   public string tournamentName { get; set; } = String.Empty;
   public string tournamentStatus { get; set; } = String.Empty;
-  public int prizeMoney { get; set; }
+  public decimal prizeMoney { get; set; }
   public int userId { get; set; }
 
 }
